Move hotspot selector to separate component
